Refactored login function
