+++ pages/createAcc.html
@@ -10,7 +10,6 @@
     <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.min.js" defer></script>
     <script src="https://kit.fontawesome.com/d527618ed6.js" crossorigin="anonymous"></script>
     <script src="/src/js/index.js" type="module" defer></script>
-    <script src="/src/js/api/services/register.js" type="module" defer></script>
     <title>TheSocial | Create Account</title>
   </head>
   <body class="bg-secondary">
@@ -55,8 +54,8 @@
         <!-- User register form -->
         <form class="regular-calibri fs-5 m-auto mb-5 w-75 was-validated" id="registerForm" action="../pages/login.html">
           <div class="mb-3">
-            <label for="username" class="form-label">Username</label>
-            <input type="text" name="username" class="form-control" id="registerUsername" minlength="8" required />
+            <label for="name" class="form-label">Username</label>
+            <input type="text" name="name" class="form-control" id="registerUsername" minlength="8" required />
           </div>
           <div class="mb-3">
             <label for="email" class="form-label">Email address</label>
