--- src/js/api/helpers/variables.mjs
@@ -1,1 +1,0 @@
-export const feedbackMsg = document.querySelector("#feedback");
